requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
sqlalchemy
pydantic-settings
tenacity
pytest
httpx
trafilatura
